Edits to Visual Studio Projects and a fix to some macros in libs/afio/src/async_file_io.cpp
This commit focused on getting VS2012 projects up to date.

An issue arose with BOOST_AFIO_POSIX_STAT macro under windows
There are "_stat64" and "_wstat64" on windows, while only one "stat" under
POSIX, so to fix the issue I changed the macros to reflect this as best as
possible.

This solution leaves something to be desired, and should be fixed with a
better name when possible.

Compiles under MSVC 2012 w/ November2012 CTP compiler