Intial step in converting from scons to Boost.Build
Added jamfiles for afio, NiallsCPP11utilities, and a jamroot for the whole project

NiallsCPP_jamfile.jam is the jamfile (normally named jamfile.jam) that resides in NiallsCPP11Utilities folder. Because it is a submodule, I cannot add this file, so for visability, and project history, I'm including this version. To use it, rename it jamfile.jam, and place it in the NiallsCPP11Utilities folder. Then run b2 from the main boost.afio directory. Note that only the libs compile at this point, and not the unit test executable.

Note that only gcc w/ c++11 is supported at this time.  Furthur work will be needed to improve compatability and supported architectures.

TODO:
1. get unit tests to build with this system
2. add full featured support for different architectures.