Successfully built unit test!
Added linkage to boost libs for the unit test executable

This is the most bare bones execution possible. There are no checks for dynamic or static linking, no compilers supported aside from gcc w/ c++11, no checks for architectures, etc. A ton more work is required for this to meet the standard.

TODO:
1. Fix the way boost is linked to the exe, we probably shouldn't be using link flags for that (eventually we'll just reference the target lib in its jamfile)
2. Add support for other architectures
3. Add relevent compiler flags (I've only done the minimum in debug build)
4. Fix the target names
5. Figure out how to build and run the unit tests with bjam, simmilar to other libs
6. Finish removing external dependencies (should be easier now that boost build works, as we can just provide the targets)