Changed BUILDPATH to TESTPREFIX & tried to make sure FLAGS is correct
BUILDPATH was a bit misleading, TESTPREFIX is more accurate, as it really is just the first part of the path to the unit test exe rather than the actual build path

I was concerned that FLAGS needed an extra set of "" to be correct. This change reflects that so the string appended in the build order is correct. This may be unneccesary, and is lilely unrelated to the current issue with clang and other build problems.