Significantly cut NiallsCPP11Utilities.hpp
Trying to completely remove NiallsCPP11Utilities.hpp

Int128_256 has some hold up with removing it and switching to Aligned_allocator.hpp. It likely has to do with the STL overrides, but my preliminary attempts to replace those references with boost::afio::detail counterparts met with failure. This is very close to running, so I'd like to fix this, and then see if we can further strip Int128_256 by relying on its SHA256 c implementation.