Build AFIO as a part of boost proper
Try to fix paths, and build calls in travis.yml

This will likely be able to build AFIO, but will probably fail on the unit tests. We'll need to refactor it to work with the boost regression testing. We may also need to set a --prefix path on the bootstrap or install step