Explicitly link libs to the test exe
Previously this was not required, as afio required these libs to be built, and was linked to them, so when the test depended on afio, it was taken care of. We'll need to examine why this is the case as we move forward. I wouldn't want the user to need to link to all these libs explicitly. This works locally for me with both static and shared libs.