Replaced Int128_256 with SpookyHash
After a lot of headaches I finally realized that memhashes was actually storing the hashed values for later comparison. I'm not sure how I failed to understand that for so long, but afterwords I was able to easily replace the Int128_256 code with direct calls to SpookyHash.

Some code clean up and additional comments are likely needed, and will be added after this passes on travis, and Niall thinks what I've done is reasonably correct.