Fix the conditional statment again
I'm not exactly sure whats wrong with this, my tests locally haven't had any issues, and the syntax has been identical. I must be makeing a mistake somehow. :(

Regardless, this test is unimportant, and I need to address how bjam is calling g++. Maybe in the user-config.jam (if it will work this time)