Changes to Jamfiles and test_all.cpp to build properly
I had improperly linked one of the tests, so it failed

We should consider looking into a way to consolidate the test results at the end of the test run, rather that having pass/fail dispersed through the entire output. Maybe there is a facility in Boost.Build that can do that.