Fixed the jamfile paths to target libs, and fixed macros for constexpr
Changed the paths to other library targets for afio to match the way other libs had done it. The previous way wasn't wrong, and might even be preferable as it doesn't rely on a releative path, but w're gaining consistency with the rest of Boost.

I also changed the macro BOOST_NO_CONSTEXPR_or_CONST to BOOST_CONSTEXPR, as falling back on const wasn't the right decision in this case.