Tried to change some std::atomic to boost::afio::atomic overload
After making an atomic that we could use with older compilers, we still
needed to us it in the unit tests. There was some namespace overlap, and
atomic<T> was ambiguous, so I specifed the namespace boost::afio:: for all
occurances of atomic.

This should be ok, as boost::afio::atomic inherits from either
boost::atomic, or std::atomic, and esentially does nothing