directory_handle::read() now holds an internal spinlock, as it is racy for multiple kernel threads to call this operation.
On Windows directory_handle::read() now has a very different implementation to before. We now do the enumerate twice, first time to calculate the kernel buffer needed for the second snapshot enumeration. The cause is that newer Windows perform ProbeForWrite() on the kernel buffer, which is a slow call with complexity quadratic to the size of the kernel buffer passed. We therefore want the smallest possible kernel buffer, even if we have to do a ton load of extra syscalls (yes, it's *that* slow).