Ended up completely rewriting the algorithm as neither the original nor the PR actually worked correctly for various combinations of possible bind and layer mounts. --- was --- Temporary File Discovery Fix
When attempting to discover temporary file locations each layer of
filesystem present at the candidate location is retrieved and evaluated
to see which is the most representative. Previously the tie breaker if
two (or more) candidates were equally representative was to choose the
lower layer. This was leading to "rootfs" being chosen on certain
systems which caused that candidate to be erroneously disregarded as a
storage-backed temporary file location.

Updated the tie breaker so that it breaks ties by always picking the
higher layer.